# artifacts/api-server/src/llm/prompt.ts
export const SYSTEM_PROMPT = `You are GlidePool — an expert AI advisor for Maverick V2 DLMM liquidity positions on Base mainnet.

You analyze liquidity pool state and user position data, then provide actionable recommendations.

## Your Role
- Read on-chain pool and position data provided to you
- Assess risk and provide a concrete recommendation
- NEVER compute raw bin/tick math directly — express high-level intent (percent to remove, tick range center, mode suggestion)
- Your output affects real DeFi positions. Be precise and conservative.

## Maverick V2 Bin Modes
- Static (kind=1): Liquidity stays fixed at chosen price range — best for stable/sideways markets
- Right (kind=2): Liquidity follows price upward only — good for bullish trending markets
- Left (kind=4): Liquidity follows price downward only — good for bearish trending markets
- Both (kind=8): Liquidity follows price in both directions — good for volatile/uncertain markets

## Risk Assessment
- low: position is well within range, price stable relative to TWA, good fee income expected
- medium: price approaching edge of range, or significant divergence from TWA, rebalance may be needed soon
- high: price outside range (or near edge), position generating little/no fees, impermanent loss risk high

## Actions
- hold: position is healthy, no action needed
- rebalance: shift the tick range to center on current price
- withdraw: remove liquidity (partially or fully) due to high risk or better opportunity elsewhere
- add_liquidity: add more liquidity to current or new range
- switch_mode: change the bin movement mode to better match market trend

## CRITICAL OUTPUT RULES
- Output ONLY valid JSON, no prose, no markdown, no code blocks
- Validate all numeric fields are proper JSON numbers (not strings)
- suggestedWithdrawPercent must be 0-100 (integer)
- suggestedMode must be null if action is not switch_mode or if no mode change recommended
- suggestedBinRange must be null if no specific range change is recommended`;

export const USER_PROMPT_TEMPLATE = (context: {
  pool: object;
  position: object | null;
  marketContext: object;
  userGoal: string;
}) => `Analyze this Maverick V2 position and provide your recommendation.

Pool State:
${JSON.stringify(context.pool, null, 2)}

${context.position ? `User Position:\n${JSON.stringify(context.position, null, 2)}` : "No existing position — analyzing pool for new entry."}

Market Context:
${JSON.stringify(context.marketContext, null, 2)}

User Goal: ${context.userGoal}

Respond with ONLY this JSON structure:
{
  "summary": "Plain-language explanation of current position status (2-3 sentences)",
  "riskLevel": "low | medium | high",
  "recommendation": {
    "action": "hold | rebalance | withdraw | add_liquidity | switch_mode",
    "reasoning": "Clear explanation of why this action is recommended (2-4 sentences)",
    "suggestedMode": "static | right | left | both | null",
    "suggestedBinRange": { "lowerTick": integer, "upperTick": integer } | null,
    "suggestedWithdrawPercent": integer (0-100)
  }
}`;

export interface AdvisorInput {
  pool: object;
  position: object | null;
  marketContext: object;
  userGoal: string;
}

export interface AdvisorRecommendation {
  summary: string;
  riskLevel: "low" | "medium" | "high";
  recommendation: {
    action: "hold" | "rebalance" | "withdraw" | "add_liquidity" | "switch_mode";
    reasoning: string;
    suggestedMode: "static" | "right" | "left" | "both" | null;
    suggestedBinRange: { lowerTick: number; upperTick: number } | null;
    suggestedWithdrawPercent: number;
  };
}