# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GlidePool – AI advisor for Maverick V2 DLMM positions on Base
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pools
    description: Maverick V2 pool data
  - name: positions
    description: User LP positions (NFT-based)
  - name: advisor
    description: LLM advisor endpoint (x402-gated)
  - name: liquidity
    description: Liquidity parameter computation

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /pools:
    get:
      operationId: listPools
      tags: [pools]
      summary: List all supported pools
      responses:
        "200":
          description: List of supported pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PoolSummary"

  /pools/{poolAddress}:
    get:
      operationId: getPool
      tags: [pools]
      summary: Get full pool state
      parameters:
        - name: poolAddress
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pool state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolState"
        "404":
          description: Pool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /positions/{userAddress}:
    get:
      operationId: getUserPositions
      tags: [positions]
      summary: Get all LP positions for a wallet address
      parameters:
        - name: userAddress
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPosition"
        "400":
          description: Invalid address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /positions/{userAddress}/{nftId}:
    get:
      operationId: getPositionDetail
      tags: [positions]
      summary: Get detailed info for a specific position NFT
      parameters:
        - name: userAddress
          in: path
          required: true
          schema:
            type: string
        - name: nftId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Position detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionDetail"
        "404":
          description: Position not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /advisor:
    get:
      operationId: getAdvice
      tags: [advisor]
      summary: Get AI recommendation for a position (x402-gated)
      parameters:
        - name: poolAddress
          in: query
          required: true
          schema:
            type: string
        - name: nftId
          in: query
          required: false
          schema:
            type: string
        - name: userGoal
          in: query
          required: false
          schema:
            type: string
            description: e.g. 'maximize fee income while limiting impermanent loss'
      responses:
        "200":
          description: AI recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvisorRecommendation"
        "402":
          description: Payment required (x402)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequired"
        "400":
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /liquidity/remove-params:
    post:
      operationId: computeRemoveParams
      tags: [liquidity]
      summary: Compute RemoveLiquidityParams from position + percent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveParamsInput"
      responses:
        "200":
          description: Ready-to-sign remove liquidity params
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoveLiquidityResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /liquidity/add-params:
    post:
      operationId: computeAddParams
      tags: [liquidity]
      summary: Compute AddLiquidityParams via PoolLens view call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddParamsInput"
      responses:
        "200":
          description: Ready-to-sign add liquidity params
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddLiquidityResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    PoolSummary:
      type: object
      required: [poolAddress, tokenA, tokenB, tokenASymbol, tokenBSymbol, tvlUsd]
      properties:
        poolAddress:
          type: string
        tokenA:
          type: string
        tokenB:
          type: string
        tokenASymbol:
          type: string
        tokenBSymbol:
          type: string
        tvlUsd:
          type: number
        activeTick:
          type: string
        currentPrice:
          type: number
        feeRate:
          type: number

    PoolState:
      type: object
      required: [poolAddress, tokenA, tokenB, tokenASymbol, tokenBSymbol, activeTick, sqrtPrice, reserveA, reserveB, feeAIn, feeBIn, tickSpacing, binCounter, lastTwaD8, currentPrice]
      properties:
        poolAddress:
          type: string
        tokenA:
          type: string
        tokenB:
          type: string
        tokenASymbol:
          type: string
        tokenBSymbol:
          type: string
        activeTick:
          type: string
        sqrtPrice:
          type: string
        reserveA:
          type: string
        reserveB:
          type: string
        feeAIn:
          type: string
        feeBIn:
          type: string
        tickSpacing:
          type: string
        binCounter:
          type: string
        lastTwaD8:
          type: string
        currentPrice:
          type: number

    UserPosition:
      type: object
      required: [nftId, poolAddress, tokenASymbol, tokenBSymbol, amountA, amountB, binCount, valueUsd]
      properties:
        nftId:
          type: string
        poolAddress:
          type: string
        tokenASymbol:
          type: string
        tokenBSymbol:
          type: string
        amountA:
          type: string
        amountB:
          type: string
        binCount:
          type: integer
        valueUsd:
          type: number

    PositionDetail:
      type: object
      required: [nftId, poolAddress, tokenA, tokenB, tokenASymbol, tokenBSymbol, binIds, amountA, amountB, valueUsd, poolState]
      properties:
        nftId:
          type: string
        poolAddress:
          type: string
        tokenA:
          type: string
        tokenB:
          type: string
        tokenASymbol:
          type: string
        tokenBSymbol:
          type: string
        binIds:
          type: array
          items:
            type: string
        amountA:
          type: string
        amountB:
          type: string
        valueUsd:
          type: number
        poolState:
          $ref: "#/components/schemas/PoolState"

    AdvisorRecommendation:
      type: object
      required: [summary, riskLevel, recommendation]
      properties:
        summary:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        recommendation:
          type: object
          required: [action, reasoning]
          properties:
            action:
              type: string
              enum: [hold, rebalance, withdraw, add_liquidity, switch_mode]
            reasoning:
              type: string
            suggestedMode:
              type: ["string", "null"]
              enum: [static, right, left, both, null]
            suggestedBinRange:
              type: ["object", "null"]
              properties:
                lowerTick:
                  type: integer
                upperTick:
                  type: integer
            suggestedWithdrawPercent:
              type: integer

    PaymentRequired:
      type: object
      required: [message, amount, currency, recipient, network]
      properties:
        message:
          type: string
        amount:
          type: string
        currency:
          type: string
        recipient:
          type: string
        network:
          type: string

    RemoveParamsInput:
      type: object
      required: [nftId, userAddress, poolAddress, withdrawPercent]
      properties:
        nftId:
          type: string
        userAddress:
          type: string
        poolAddress:
          type: string
        withdrawPercent:
          type: integer
          minimum: 1
          maximum: 100

    RemoveLiquidityResult:
      type: object
      required: [poolAddress, recipient, subaccount, binIds, amounts, estimatedTokenA, estimatedTokenB]
      properties:
        poolAddress:
          type: string
        recipient:
          type: string
        subaccount:
          type: integer
        binIds:
          type: array
          items:
            type: string
        amounts:
          type: array
          items:
            type: string
        estimatedTokenA:
          type: string
        estimatedTokenB:
          type: string

    AddParamsInput:
      type: object
      required: [poolAddress, kind, lowerTick, upperTick, userAddress]
      properties:
        poolAddress:
          type: string
        kind:
          type: integer
          enum: [1, 2, 4, 8]
          description: "1=Static, 2=Right, 4=Left, 8=Both"
        lowerTick:
          type: integer
        upperTick:
          type: integer
        userAddress:
          type: string
        tokenAAmount:
          type: ["string", "null"]
        tokenBAmount:
          type: ["string", "null"]

    AddLiquidityResult:
      type: object
      required: [poolAddress, kind, lowerTick, upperTick, estimatedTokenA, estimatedTokenB, params]
      properties:
        poolAddress:
          type: string
        kind:
          type: integer
        lowerTick:
          type: integer
        upperTick:
          type: integer
        estimatedTokenA:
          type: string
        estimatedTokenB:
          type: string
        params:
          type: object
          description: Encoded AddLiquidityParams for the frontend to pass to the pool contract